Latest versions added

+++ qemu_mode/build_qemu_support.sh
@@ -6,7 +6,7 @@
 # Written by Andrew Griffiths <agriffiths@google.com> and
 #            Michal Zalewski <lcamtuf@google.com>
 #
-# Copyright 2015 Google Inc. All rights reserved.
+# Copyright 2015, 2016 Google Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,8 +22,8 @@
 # will be written to ../afl-qemu-trace.
 #
 
-QEMU_URL="http://wiki.qemu-project.org/download/qemu-2.2.0.tar.bz2"
-QEMU_SHA384="69f4ac3094b0577b7181840c9c7b7a048df2bd03c0d851eef8174fd052a1ba786cff15a7dbd94410cbfcb53cb823a30c"
+QEMU_URL="http://wiki.qemu-project.org/download/qemu-2.3.0.tar.bz2"
+QEMU_SHA384="7a0f0c900f7e2048463cc32ff3e904965ab466c8428847400a0f2dcfe458108a68012c4fddb2a7e7c822b4fd1a49639b"
 
 echo "================================================="
 echo "AFL binary-only instrumentation QEMU build script"
@@ -89,7 +89,7 @@ CKSUM=`sha384sum -- "$ARCHIVE" 2>/dev/null | cut -d' ' -f1`
 
 if [ ! "$CKSUM" = "$QEMU_SHA384" ]; then
 
-  echo "[*] Downloading QEMU 2.2.0 from the web..."
+  echo "[*] Downloading QEMU 2.3.0 from the web..."
   rm -f "$ARCHIVE"
   wget -O "$ARCHIVE" -- "$QEMU_URL" || exit 1
 
@@ -110,7 +110,7 @@ fi
 
 echo "[*] Uncompressing archive (this will take a while)..."
 
-rm -rf "qemu-2.2.0" || exit 1
+rm -rf "qemu-2.3.0" || exit 1
 tar xf "$ARCHIVE" || exit 1
 
 echo "[+] Unpacking successful."
@@ -124,12 +124,14 @@ patch -p0 <patches/syscall.diff || exit 1
 
 echo "[+] Patching done."
 
+ORIG_CPU_TARGET="$CPU_TARGET"
+
 test "$CPU_TARGET" = "" && CPU_TARGET="`uname -m`"
 test "$CPU_TARGET" = "i686" && CPU_TARGET="i386"
 
 echo "[*] Configuring QEMU for $CPU_TARGET..."
 
-cd qemu-2.2.0 || exit 1
+cd qemu-2.3.0 || exit 1
 
 CFLAGS="-O3" ./configure --disable-system --enable-linux-user \
   --enable-guest-base --disable-gtk --disable-sdl --disable-vnc \
@@ -152,35 +154,43 @@ ls -l ../afl-qemu-trace || exit 1
 
 echo "[+] Successfully created '../afl-qemu-trace'."
 
-echo "[*] Testing the build..."
+if [ "$ORIG_CPU_TARGET" = "" ]; then
 
-cd ..
+  echo "[*] Testing the build..."
 
-make >/dev/null || exit 1
+  cd ..
 
-gcc test-instr.c -o test-instr || exit 1
+  make >/dev/null || exit 1
 
-unset AFL_INST_RATIO
+  gcc test-instr.c -o test-instr || exit 1
 
-echo 0 | ./afl-showmap -m none -Q -q -o .test-instr0 ./test-instr || exit 1
-echo 1 | ./afl-showmap -m none -Q -q -o .test-instr1 ./test-instr || exit 1
+  unset AFL_INST_RATIO
 
-rm -f test-instr
+  echo 0 | ./afl-showmap -m none -Q -q -o .test-instr0 ./test-instr || exit 1
+  echo 1 | ./afl-showmap -m none -Q -q -o .test-instr1 ./test-instr || exit 1
 
-cmp -s .test-instr0 .test-instr1
-DR="$?"
+  rm -f test-instr
 
-rm -f .test-instr0 .test-instr1
+  cmp -s .test-instr0 .test-instr1
+  DR="$?"
 
-if [ "$DR" = "0" ]; then
+  rm -f .test-instr0 .test-instr1
 
-  echo "[-] Error: afl-qemu-trace instrumentation doesn't seem to work!"
-  exit 1
+  if [ "$DR" = "0" ]; then
+
+    echo "[-] Error: afl-qemu-trace instrumentation doesn't seem to work!"
+    exit 1
+
+  fi
+
+  echo "[+] Instrumentation tests passed. "
+  echo "[+] All set, you can now use the -Q mode in afl-fuzz!"
+
+else
+
+  echo "[!] Note: can't test instrumentation when CPU_TARGET set."
+  echo "[+] All set, you can now (hopefully) use the -Q mode in afl-fuzz!"
 
 fi
-
-echo "[+] Instrumentation tests passed. "
-
-echo "[+] All set, you can now use the -Q mode in afl-fuzz!"
 
 exit 0
